Rename all components to kebab-case
