+++ src/app/[...slug]/page.tsx
@@ -1,6 +1,6 @@
 import { notFound } from 'next/navigation';
 
-import { Mdx } from '@/components/layout/Mdx';
+import { Mdx } from '@/components/layout/mdx';
 import { allDocs, specialFileNames } from '@/constants/contentlayer';
 
 export const generateStaticParams = async () =>
