--- src/components/general/Collapsible/index.ts
@@ -1,1 +1,0 @@
-export * from './Collapsible';
