--- src/components/general/List/index.ts
@@ -1,5 +1,0 @@
-export * from './List';
-export * from './ListItem';
-export * from './ListItemIcon';
-export * from './ListItemText';
-export * from './ListSubheader';
