--- src/components/general/Sheet/index.ts
@@ -1,1 +1,0 @@
-export * from './Sheet';
