+++ src/components/general/collapsible/index.ts
@@ -1,0 +1,1 @@
+export * from './collapsible';
