+++ src/components/general/index.ts
@@ -1,5 +1,5 @@
-export * from './Collapsible';
-export * from './IconButton';
-export * from './List';
-export * from './ScrollArea';
-export * from './Sheet';
+export * from './collapsible';
+export * from './icon-button';
+export * from './list';
+export * from './scroll-area';
+export * from './sheet';
