+++ src/components/general/scroll-area/index.ts
@@ -1,0 +1,2 @@
+export * from './scroll-area';
+export * from './scroll-bar';
