--- src/components/layout/Mdx/index.ts
@@ -1,4 +1,0 @@
-export * from './Alert';
-export * from './Mdx';
-export * from './MdxBlockquote';
-export * from './MdxLink';
