--- src/components/layout/Nav/index.ts
@@ -1,3 +1,0 @@
-export * from './Nav';
-export * from './NavLink';
-export * from './NavLinks';
