--- src/components/layout/Sidebar/index.ts
@@ -1,2 +1,0 @@
-export * from './Sidebar';
-export * from './SidebarItem';
