+++ src/components/layout/mdx/index.ts
@@ -1,0 +1,4 @@
+export * from './alert';
+export * from './mdx';
+export * from './mdx-blockquote';
+export * from './mdx-link';
