+++ src/components/layout/search/index.ts
@@ -1,0 +1,1 @@
+export * from './search';
