+++ src/components/layout/sidebar/index.ts
@@ -1,0 +1,2 @@
+export * from './sidebar';
+export * from './sidebar-item';
